backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fixwise-backend"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
    "openai>=1.50",
    "Pillow>=11.0",
    "cryptography>=44.0",
    "PyJWT>=2.10",
    "redis>=5.0",
    "SQLAlchemy>=2.0",
    "asyncpg>=0.30",
    "boto3>=1.35",
    "pydantic[email]>=2.10",
    "python-multipart>=0.0.20",
    "httpx>=0.28",
    "aiosqlite>=0.20",
    "reportlab>=4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
